extract wasm_ql_http
